restructure and linting
