+++ client/.eslintrc.json
@@ -1,0 +1,116 @@
+{
+    "extends": ["airbnb"],
+    "parser": "babel-eslint",
+    "parserOptions": {
+      "ecmaVersion": 8,
+      "ecmaFeatures": {
+        "experimentalObjectRestSpread": true,
+        "impliedStrict": true,
+        "classes": true
+      }
+    },
+    "env": {
+      "browser": true,
+      "node": true,
+      "jquery": true,
+      "jest": true
+    },
+    "rules": {
+      "no-debugger": 0,
+      "no-unused-vars": [
+        1,
+        {
+          "argsIgnorePattern": "res|next|^err"
+        }
+      ],
+      "prefer-const": [
+        "error",
+        {
+          "destructuring": "all"
+        }
+      ],
+      "arrow-parens": ["error", "always"],
+      "arrow-body-style": [
+        2,
+        "as-needed"
+      ],
+      "jsx-a11y/no-autofocus": 0,
+      "jsx-a11y/click-events-have-key-events": 0,
+      "no-unused-expressions": [
+        2,
+        {
+          "allowTaggedTemplates": true
+        }
+      ],
+      "no-param-reassign": [
+        2,
+        {
+          "props": false
+        }
+      ],
+      "no-console": 2,
+      "import/prefer-default-export": 0,
+      "import": 0,
+      "func-names": 0,
+      "space-before-function-paren": 0,
+      "comma-dangle": ["error", {
+        "arrays": "always-multiline",
+        "objects": "always-multiline",
+        "imports": "always-multiline",
+        "exports": "always-multiline",
+        "functions": "ignore"
+      }],
+      "max-len": 0,
+      "import/extensions": 0,
+      "no-underscore-dangle": 0,
+      "consistent-return": 0,
+      "react/display-name": 1,
+      "react/no-array-index-key": 0,
+      "react/react-in-jsx-scope": 0,
+      "react/prefer-stateless-function": 0,
+      "react/forbid-prop-types": 0,
+      "react/no-unescaped-entities": 0,
+      "jsx-a11y/accessible-emoji": 0,
+      "react/jsx-one-expression-per-line": 0,
+      "react/jsx-filename-extension": [
+        1,
+        {
+          "extensions": [
+            ".js",
+            ".jsx"
+          ]
+        }
+      ],
+      "radix": 0,
+      "no-shadow": [
+        2,
+        {
+          "hoist": "all",
+          "allow": [
+            "resolve",
+            "reject",
+            "done",
+            "next",
+            "err",
+            "error"
+          ]
+        }
+      ],
+      "quotes": [
+        2,
+        "single",
+        {
+          "avoidEscape": true
+        }
+      ],
+      "jsx-a11y/href-no-hash": "off",
+      "jsx-a11y/anchor-is-valid": [
+        "warn",
+        {
+          "aspects": [
+            "invalidHref"
+          ]
+        }
+      ]
+    }
+  }
